Fix a problem with the "once" function.
Formerly this program would print "Hi" three times instead of only once:
~~
\hi=(once; say "Hi")
I  # This was the problem.
hi
hi
hi
~~

The problem was that (once x) was replacing the entire (once x) form with the
value of x.  That was normally fine, but calling the "I" function broke that
behavior because it makes a shallow copy of its argument.  The copy would run
and be replaced with its final value "I".  Then that "I" value would make a
copy of the next occurrence of the "hi" function, and it too would run, all the
way down the line.

I fixed that by making (once x) only replace x with its final value, instead of
trying to replace the entire (once x) form.

I also made a cosmetic change to type_rand.