Dramatically simplify and accelerate symbol resolution.
Formerly, the resolve function was laboriously finding the next symbol in the
expression, abstracting it out using the lambda mechanism, and then applying
that to the symbol's definition.  This, I found, was a serious problem with a
test case I created that contains 31,251 symbols.  It churned for about 90
seconds and used a ton of memory.

Not anymore.  The new resolve function simply runs through the expression
recursively, replacing each occurrence of a symbol with its definition right
then and there.  Now the big test case with 31,251 symbols runs in 0.15 seconds
and uses only a tiny amount of memory.

With these changes, if you have undefined symbols in your code, Fexl reports
every occurrence of that symbol as a separate error message, instead of only
one error message per unique symbol.  However, that is just fine with me,
given the simplicity and speed of the new resolve routine.  Even a C compiler
will give you a cascade of error messages for undefined symbols, so there's no
sense in me going out of my way to keep them unique in Fexl.

The new code is not only far faster, but the executable is also 1120 bytes
smaller.

I also removed a couple of stray comments in test/a1.fxl.