Make the "slice" function more convenient.
Formerly if pos or len extended beyond the bounds of the string in *any* way,
it would return void.  That is very inconvenient in a case like this:

	(slice "abc" 3 0)

Formerly that returned void because position 3 (counting from 0) was just past
the end of the string -- even though the desired length was 0, so the operation
is harmless.  Now it more conveniently returns "" (the null string).

That's great, but I went further.  The slice function now conveniently clips
any range that exceeds the bounds of the string.  Therefore all of these are
now true instead of void:

	eq (slice "abc" 0 4) "abc"
	eq (slice "abc" 1 3) "bc"
	eq (slice "abc" 2 9) "c"
	eq (slice "abc" 8 2) ""
	eq (slice "abc" 8 0) ""

As an example of why this is convenient, here is how you test if a string
starts with "xyz":

	eq (slice x 0 3) "xyz"

You no longer have to do a separate test to handle the case where x has fewer
than 3 characters.

I also extended and reordered the slice tests in test/a1.fxl.