Now using long-reach substitution patterns.
The efficiency gains in this release are fairly small: test/fib runs about 2%
faster, and test/checkt runs about 0.5% faster.

However, I think this is just the beginning.  Now that we're using C function
pointers in the patterns, we should be able to optimize common substitution
patterns by implementing new builder functions and using them in fuse.

Also, for what it's worth, the new executable is 780 bytes smaller, which
probably reflects the simplification of the subst routine, which now simply
calls a function pointer instead of doing pattern matching.