Eliminate attempt at "long-reach" substitution.
I'm now going back to simple substitution based on examination of the pattern
at run-time, during the call to "subst", instead of using "builder" function
pointers.

Amazingly, this code is now about 4% faster than what I had before.  It seems
clear that using the function pointers did not yield any actual benefits.

I thought that the function pointer approach would enable deeper optimizations
based on noticing certain very common substitution patterns, but I played
around with that for quite a while and things just got more complex and even
slightly slower.  I tried matching common bottom-up patterns, and I tried
matching common top-down patterns, to no avail.  Those results seem to be
telling me that this is the wrong place to attempt optimization.  It seems that
at some fundamental level, substitution is a sort of irreducible primary that
does not yield easily to optimization.

On further reflection, I think the better approach will be to write certain
highly useful functions such as "each", "range", "fold", "filter", and others
directly in C, instead of trying to optimize "substitution in general."

I know in the past that has yielded excellent results, because in an older
version of the code I once wrote "fold" directly in C, and certain tests became
70% faster right away.  It will be interesting to see if, for example, writing
"filter" in C will accelerate the sort.fxl test.

Ultimately it might be possible to write a sort of "Fexl compiler", in the
sense that it would translate each substitution pattern it encountered into
static C code that did the substitution "all at once" without any recursive
calls to the generic "subst" routine in pattern.c.  However, that may not even
be desirable, since I suspect that just rewriting a few specific useful
functions in C will yield the greatest gains without going to the trouble of
generating C code.  Besides, Fexl needs to be interpreted, because it's my
web scripting language and I'm certainly not going to fire off a "gcc" process
every time I change a calculation or a table format.

Interestingly, at this point the fib.fxl test takes 0.28s, while an equivalent
version written in Perl takes 0.14s.  So the Fexl version is twice as slow --
but that's pretty darn good considering everything that's going on in Fexl.
Perhaps if I wrote "each" and "range" in C, the Fexl version could be just as
fast.

Don't get me wrong here, I know that sheer speed isn't everything.  If it was,
I would just write all my application code directly in C.  However, that would
not give me the platform I need to have a powerful sandboxed scripting language
to drive our accounting web sites, and it would be tedious and error-prone
besides.  The main reason I occasionally go on a quest for more speed is to
prove to myself that the language can scale up to meet any demands I may put
upon it in the future.  The ability to wring out an extra 20% in speed here and
there is not currently mission-critical, but it's nice to know I can do it if
necessary.

To reiterate my thoughts above, I think the best opportunities for optimization
will be to rewrite common control structure functions in C, and not worry too
much about generic optimization of the interpreter, which seems to have reached
a fundamental limit of efficiency anyway.  In the meantime, I'm quite happy
with the flexibility, reliability, expressiveness, and speed of the language
for real work.