Fix special case in resolve.
Formerly when I resolved a symbol in dynamic_context I would compare the left
side for equality with a token called "single".  I devised a test case that
is logically valid but fails the equality test.  Now I compare the type of the
left side for equality with type_single, and that fixes the test case.  This is
a more rigorous approach for testing the form of an evaluated expression within
C code.