Changed the signatures of the parse functions.
Formerly parse_file and parse_string returned the value directly and replaced
itself inline.  That made it impossible to parse a file repeatedly and get a
different result if the file changed or the file was a stream such as stdin.
It also made it impossible to parse a file or string and get the function
without actually evaluating it.

Now the parse functions return a singleton {fn}, where fn is the function, so
you can use them like this:

	parse_file name context \fn ...

	parse_string str context \fn ...

I bumped the major version number to 3 because this change is not compatible
with version 2.