Use "arg" routine to evaluate all args in place.
This simplifies a lot of code, so instead of saying this:

	value x = eval(hold(f->R));
	...
	drop(x);

I can just say this:

	value x = arg(f->R);
	...

Note that I don't have to do the final drop.

I also renamed all "replace*" routines to "reduce*".

I added some tests in a1.fxl to further nail down the behavior of embedded
side effects.  Note that the behavior of the test labeled "normal definition"
changes slightly now, since the argument of say is always evaluated in place.
This is technically a backward incompatibility, but I didn't think it merited
a bump in the major version because it's so subtle.

The behavior of fopen is a bit more convenient in the case where the open
fails.  Formerly the entire form (fopen path mode) would return void.  Now it
simply returns a void file handle, which is easier to check.