Eliminate "==" for eager evaluation and use it for recursive definitions.
I know I've flip-flopped on that a couple times, and for that I'm sorry, but
it's a long story.

This version is not backward-compatible, which is why I bumped the major
version to 5.

I eliminated the "==" for eager evaluation, which is not particularly useful.
Instead, I introduced an "eval" routine for those rare occasions when you want
to force an evaluation.  This eliminated type_query and introduced type_eval
instead.

That gave me the opportunity to restore the "==" syntax to designate recursive
definitions, instead of using the clunky-looking "=\".

I eliminated the entire notion of a "current output" in output.c.  That was not
a robust concept.  I added a test in test/a1.fxl (called "test_2") which has an
embedded "say" inside a call to "fput".  The correct behavior is for the output
of say always to go to stdout, and not into the file, and this is what happens
now.  Formerly, because of the "current output" mechanism, the output of say
also went into the file, which is not correct.

I eliminated "put_to_error", which relied on the "current output" concept now
considered harmful, and added a simple "error" function for sending output to
stderr.

Added test/a3.fxl to demonstrate that "error" really does print to stderr.

Added test/hailstone.fxl just because I had it.