Accelerate some Fexl functions.
Most notably I rewrote the "filter" function in test/lib.fxl so it doesn't
have to pass in the value of "f" to each recursive call.  Instead, the "f"
value is baked into an internal "filter" function which only passes the current
list to each call.

That one change caused the test/sort.fxl test to run 37% faster.

I went ahead and applied the same technique to other recursive functions such
as "each", "map", etc.

I also added a test suite for the format_num routine.

I eliminated the explicit use of the fixpoint operator (@) in the "extend"
function in test/lib.fxl, in favor of "==" to define recursion.  I think the
"==" is a little more clear.

I simplified and optimized test/mf.fxl, and it runs 37% faster.