Fix bug in type_var_put.
Formerly it would drop the old value first, then evaluate the new value and
store it.  I created a case where the new value itself used var_get to get the
old value.  That caused a fatal segmentation fault because the old value had
already been recycled, so the var_get returned a pointer to garbage.

This is fixed now, and I've added a test case for it.