Fix a problem in fexl_quote.
The fexl_quote function in test/lib.fxl was doing something wrong when given
this string:

	~ "a b"~

It figured it could wrap that in "~~" terminators, since "~~" does not appear
within the string.  The result was:

	~~ ~ "a b"~~~

However, that's not good, because if you parse it back, the "~~" appears first
where you don't expect it:

	~~ ~ "a b"~~~
	          ^^

It's best to avoid using "~" inside a terminator string.  Sometimes it's OK,
but it can cause a problem if the content string ends in "~".  So I changed
fexl_quote to use "~1" instead of "~" for the first terminator candidate after
plain quote marks.  The result of quoting the string is now:

	~1 ~ "a b"~~1