Eliminate inline reduction in the C routine "arg".
Formerly "arg" would always reduce its argument to its final form, in place.
I don't like that kind of behavior, because it can destroy functions which are
otherwise repeatable.  There is also no way to express that kind of behavior
in Fexl itself, which is good.

This change affected the "once" function, and I've modified a few tests to
work around the slight difference in behavior.  This change exposes a
fundamental problem I've been noticing lately.  The type_I routine calls
"reduce" to replace (I x) directly with the value of x.  The problem is, that
copies x up into the parent form, which has subtle effects on behavior.  You
can see its effect on the behavior of "once", but you can also see its effect
on more conventional evaluation, including a "run" test which I've added to
test/a1.fxl.

I am working on a version of type_I which does not copy its argument up like
that.  This has been a surprisingly subtle problem to solve, because it
challenges the assumption that pointer equality is synonymous with the non-
repetition (inline reduceability) of values.  My workaround so far is to use a
global "action" flag to indicate when a repeatable action occurred, instead of
detecting that by noticing when two pointers are not equal.

I also added a simple test of the flock_ and limit_ calls, just to extend the
test coverage a bit.

Incidentally, I changed some wording in comments from second-person "we" to
first-person "I".