Fix the loop 900000 in test/a1.fxl so it runs in constant space.
Formerly memory would "pile up" because type_A was holding onto its left side
throughout the entire evaluation.  I remedied this with a new routine called
"next_action" which iterates the type function only until the next action, if
any.  That gives the code a chance to drop the left side as it progresses.

I also changed (query x) to evaluate its argument immediately and yield it as
{v}, instead of waiting for a second argument.  It too uses the next_action
routine.  The rule is that any routine which needs to evaluate an argument and
determine whether or not the value is functionally equivalent to the original
should call next_action instead of arg.

In a future version I intend to obsolete the "later" concept and use "yield"
instead to return functions that are not fully evaluated, primarily in the
definition of symbol contexts.  I used to do it that way, but thought "later"
might be a good idea.  It isn't.  I have devised tests which demonstrate that
"later" is an untenable concept with fragile semantics.