Fix bug in str_map.
When I changed the string mapping functions to use readstr, I didn't notice
that the new str_map function was returning a list of strings instead of
flattening those into a single string.  I've fixed that now.

The problem was not caught in the test suite because printing a list of strings
looks the same as printing the flattened list.  I have added some tests to
catch this case now.

I renamed str_list to get_str_list to reflect that it maps "get" and "str" into
a list.