Avoid copy up in reduce.
Formerly the reduce routine would copy up the contents of g into f.  This would
break certain test cases involving the "once" function.

Now the reduce function sets f to the intermediate form (J g), which avoids the
copy of the components of g into f.  The type_J routine simply returns g.  Now
that causes a situation where the return value is functionally equivalent to
f, but not equal to f in terms of pointer equality.  That's a first, so I had
to introduce an "action" flag to indicate when a return value is not
functionally equivalent, since I could no longer rely on pointer inequality.

Also, because contents are no longer copied up, I was able to eliminate the
double-indirection in the representation of atomic data values.  In the next
revision I will eliminate the "data" routine altogether.

By eliminating the double-indirection, I also eliminated the structure which
held the "destroy" routine for atomic data.  I now handle that by calling the
type routine and having it check for a zero reference count.