Make it faster.
This version 8.2.0 is faster than 8.1.1, and uses less memory.  It is very much
like version 8.0.1, but also preserves the proper behavior of the "once" tests.

I rolled back the "action" flag used in 8.1.1.  I once again replace values
inline, just as in 8.0.1, but now that is confined to the type_J function,
which replaces the value only after it is fully evaluated.  That is the key to
preserving the proper "once" behavior.

The test/checkt runs in 2.03s in version 8.2.0, and it ran in 2.50s in version
8.1.1.  That makes 8.2.0 about 22% faster than 8.1.1.  It is essentially the
same speed as 8.0.1.