Simplify eval loop.
I eliminated the next_action concept because ultimately it is useless at
preventing the kind of out of memory error exhibited in a4.fxl.  I was using it
in type_A and type_eval, but I would have to use it in every single arg() call
to chase down every possible case, and that is way too convoluted.  The simple
workaround is a semicolon, so there's no sense in continuing that losing fight.

I added some more "once" test cases to a1.fxl.

I simplified hailstone.fxl slightly.

I simplified beer.fxl, making it look clearer.  The script is now 258 bytes
instead of 333 bytes, a 22% reduction in size.