Make "once" into a syntactic construct instead of a function.
I have eliminated the "once" function and replaced it with the "\=" syntax
instead.  This allows me to eliminate the pesky and ugly type_J function once
and for all.  Now I can do a simple "reduce" function which copies it straight
up instead of going through the type_J gyrations.

You can't do that if you use "once" as a callable function.  For example if you
have a value y=(once x) and a value z=(I y), the I function copies the the y
value up, so now z=(once x).  However, the original y value is still (once x).
So the z evaluates, and any side-effects in x occur.  However, if the y also
evaluates separately, the side-effects in x occur *again*.  Not good.  Twice is
not once.

I have a lot of test cases that push the boundaries of "once" semantics in this
way.  To pass those tests I had to introduce type_J which copies up only after
the final value is reached.

Now I can pass all those tests without type_J.  However, I have to make "once"
a syntactic construct instead of a callable function.  Now instead of saying
this:

	once x

You say this:

	\= x

(The space after the "=" is optional.)

Here's how I arrived at this solution.  First I dropped type_J entirely, using
the simple reduce function.  That broke a bunch of detailed "once" tests of
course.  Then I looked for a workaround.  I noticed that if I just doubled up
the "once" function, I could pass all the tests.  So for example where I
formerly had this:

	(once x)

I could replace it with this:

	(once (once x))

Then I thought OK, that's a workaround but it's a real pain to impose on the
user.  Clearly there is something about "once" that doesn't WANT it to be an
ordinary callable function.  Instead, it should be a special syntactic
construct.

So that's what I did.  We already had a "\;" syntax for unresolved forms, so I
extended a new similar syntax "\=" to represent the once form.  It actually
does the double-once technique inside the new parse_once function.

I bumped the major version number on this release to 10, because it is not
strictly compatible with verison 9.