Restructure to prepare for unification with new branch.
I now use a consistent scheme for ordering included header files.  They are now
in sections separated by new lines to reflect dependencies.  The rule is that
an include in one section is a prerequiste for the code that follows that
section.  Includes within a section are listed alphabetically.

I made some changes to basic.c.  I use a common op_boolean to implement type_T
and type_F.  I use a common op_predicate to implement is_good, is_bool, and
is_list.  This captures common patterns and reduces the code line count.  I
also moved op_is_type up before its first use.

In file.c I use fput_ch in the fnl routine.

In lib_build I include the -Wdeclaration-after-statement option, but pounded
out.  I unpound that when developing to ensure compatibility with ISO C90.

I added a built-in "pi" function to Fexl.  This uses the built-in M_PI from
math.h for better accuracy.  Then I removed the definition from lib.fxl.  This
extra accuracy had a visible effect in run.out, in the very low order digits of
normally distributed random numbers.

I streamlined some code in parse.c.  I'm using a version of the C routine "lam"
which requires the caller to drop its arguments.  That's because I'm about to
go with a new simpler lam routine, though not in this version.

I isolated calls to data(), putting them inside individual "get" functions such
as get_str, get_num, etc.

I added a bunch of code to show.c, which of course is not ordinarily compiled
and linked in production code, but used only for testing and development.  This
new code includes a nice evaluation routine which traces its computation.

For atom values I no longer use a "destroy" routine to free the atom data.
Instead the clear routine in value.c simply calls the type function, which now
checks for a zero reference count to indicate that it should free the atom
data.  This is clean, and is also in preparation for a new version where I no
longer using double-indirection for atoms.

In the parse_file routine I now close the file handle unless it's stdin.

I went ahead and made "eval" a named function in Fexl, even though the "=="
syntax makes it strictly unnecessary.  It could still come in handy if you're
writing code and wish to avoid "==" for some reason.

The free list code in value.c is now doing a "lazy" drop.  Before, when a
value's reference count when to zero, the code would clear the contents of the
value recursively before putting it on the free list.  Now it just pushes the
value onto the free list without clearing it.  Then, when a value is popped off
the free list, its contents are cleared at that time.  This means that "drop"
is no longer a recursive routine.

I also moved hold and drop up in the value.c file.

All of these changes reduced the executable size from 76184 to 63896 bytes, a
savings of 12288 bytes.