Report each distinct undefined symbol only once.
Now if you have several occurrences of a particular undefined symbol, only the
first occurrence is reported.  This eliminates excessive cascading error
messages.

The new report_undef routine iterates through each undefined symbol in the
form, reporting the error, and then using lam() to abstract that symbol out of
the form so it doesn't get reported again on the next pass.