Make a separate limit.c module for setting resource limits.
Add a few more tests of output functions.

Raise the loop count in a4.fxl so that it actually runs out of memory if you
pound out the semicolon.  The old count of 90000 used to do the trick, but now
it has to be 100000.

Eliminate static struct in Qsubst.  Now it just holds onto QI instead.