Use a destroy routine again in atom constructors.
This is cleaner, and made it possible to reduce a file handle to void when
calling fclose.  I added a test for this in a1.fxl, and also added a few
explicit calls to fclose in some test scripts.

Clarified the code in type_J, op_getc, and op_get.  Reordered some cases in
show.c to match header file order more closely.  Reordered some things in
basic.h and basic.c for better consistency.