Clear value before pushing onto free list.
This is what I used to do a long time ago, but at some point I thought it
clever to put values on the free list without clearing them first, and then
clear them when popping them off the free list.  Turns out that's not as clean.

Now I'm reverting to clearing the value before pushing it onto the free list.
One big advantage is that I can go back to checking for f->N == 0 in the
destructor calls instead of having to set f->T to zero and then checking
f->T == 0.  That's much simpler and cleaner.

Another advantage is that this way runs faster.  I noticed that the sort.fxl
test by itself runs 15% faster now.