Do pure evaluation, with no inline modification of values.
Evaluation is now done in a completely pure manner, with no modification of
values as they are evaluated, with the exception of type_O which implements the
"once" operator.  All "reduce" routines are gone.

The "once" operator now has a syntax of its own, which is "=;", for example:

	\y=;(f x)

That gives a value y which is not yet evaluated, but if is eventually evaluated
its value will be (f x) and it will not be evaluated again.

I was happy to discover that pure evaluation can still be made efficient.  It
does require a few changes to some Fexl code though.  For example, the sort and
unique routines are the same, but the filter routine now uses the once operator
"=;" to avoid doing repeated work.  The space-time tradeoff is now under more
direct control of the Fexl programmer.

Pure evaluation also has far more predictable memory usage behavior.  Tests
such as a4.fxl run in constant memory.  It also allows the printing of infinite
lists in constant memory, for example:

	put (@\list ["a";list])

Pure evaluation also drastically simplifies the C code.  It even reduces the
executable size to 56240 bytes from the former 76464 bytes on this machine.