Improved checking of hold/drop imbalances.
Formerly the drop routine was not recursive.  When a reference count reached
zero, it would push it on the free list.  When the value was popped off the
free list, it would be cleared at that time.

Now the drop routine clears the value before pushing it on the free list.
I also added a check in case the programmer inadvertently drops a value whose
reference count is already zero.

This change was not strictly necessary, since the code works fine either way.
However, I noticed during the development of new built-in types that there were
cases when it didn't matter whether or not I called hold or drop on a value, or
even if I made extra calls.  I didn't like that one bit because I expect the
end_value check to catch any hold/drop imbalances.  I traced though what was
going on, and found that I was dropping a value that was already on the free
list.  In that case the reference count N is a pointer value, which looks like
a large unsigned integer, and the drop proceeds somewhat normally.

Note that I only discovered this when deliberately making an error in the code.
However, I wanted to make sure accidental errors were caught as well, and this
new version does that.