Refactor.
Eliminated type_standard.c, moving all of that code directly into fexl.c.

Moved main_argc and main_argv into type_run.c.

The typedef for "buffer" now is a pointer to the struct instead of the struct
itself.  Now ALL typedefs in the code are pointer types.

Eliminated typedef of "symbol".

Eliminated typedef of "form".

Renamed the "parse" routine in parse.c as "parse_input" to make it more easily
distinguishable.

Enhanced show.c so that the show routine imposes limits on both the depth and
the breadth of the value being displayed.