Make tuples observable data types.
Formerly parse_tuple returned a raw function to implement tuples.  Now it wraps
that raw function inside a call to type_tuple to make it an observable data
type.

I added a Fexl function "is_tuple" to detect when something is a tuple, along
with a function "arity" to return the number of components in a tuple.

I enhanced the tuple tests in a1.fxl.