Check the return value from stat in is_file and is_dir.
On an older machine I noticed that when a file doesn't exist, the stat routine
would sometimes fail with return code -1, rather unpredictably I might add,
instead of just setting the status block with a usable st_mode field.  It was a
rookie mistake for me not to check the return value anyway.  Now I check it.

I also consolidated the code for type_is_file and type_is_dir into a common
routine op_stat_type which does all the necessary checking, in keeping with the
"DRY" principle (Don't Repeat Yourself).