Use type double in fexl_benchmark to handle negative memory usage.
The fexl_benchmark function was subtracting unsigned long values to compute the
number of steps and bytes used during evaluation.  In the case where memory is
freed, the number of bytes used can be negative.  To handle this case, I now
cast the values to type double before subtracting.

The number of steps can never be negative, unless somehow you run so long that
you exceed 2^64 steps and roll the counter over.  Nevertheless I handle steps
the same way as bytes.