Integrated crypto functions from tweetnacl.c
crypto.c is a "cleaned up" version of tweetnacl.c which avoids all the complex
macro definitions.  It also changes the interface somewhat.  For example, it
no longer calls an external "randombytes" function.  Instead, it requires
random material to be passed in from the caller.  In particular the key
generation functions are now redefined as functions which map a random secret
key to the corresponding public key.  It also standardizes on the "beforenm"
and "afternm" technique, requiring the user to call crypto_box_prepare to fold
a public and secret key together for use with crypto_box_seal and
crypto_box_open.

The types.h file includes the typedefs needed in the crypto library.

Because crypto.c uses "long long" types, I had to eliminate -pedantic from
lib_build because the build was failing on one of my servers.

type_hex.c has native C definitions of map_raw_bytes_to_hex and
map_hex_to_raw_bytes which were formerly defined in lib/hex.fxl.

type_crypto.c has the wrappers for the calls to the core crypto.c library.  It
defines the following functions:

	random_bytes
	random_nonce
	random_secret_key
	crypto_box_public
	crypto_box_prepare
	crypto_box_seal
	crypto_box_open
	crypto_sign_public
	crypto_sign_seal
	crypto_sign_open
	crypto_hash

For slightly enhanced security I considered calling explicit_bzero in memory.c
to zero all data being freed.  However, that call does not exist on one of my
servers so I just gave up on it for now.