Added client/server functions.
The start_server function (type_run.c) starts a server.  The stop_server
function (lib/run.fxl) stops a server.  The run_server function (lib/run.fxl)
takes care of automatically stopping and starting a server as needed, based on
an option which is either "start" or "stop".  For an example, see
test/server.fxl.

The connect function (type_run.c) connects to a server.  For an example, see
test/client.fxl.

The set_alarm function sets an alarm for a given number of seconds.  This is
used primarily by server processes so they don't wait forever for input from
the client.  If the alarm goes off it will interrupt the system call (fget,
fread, etc.) and the process can exit.

The kill function (type_run.c) sends a kill signal to a process.  This is used
in stop_server (lib/run.fxl).

Note that my client/server functions are "opinionated," meaning that they do a
specific wholistic task meant to be reliable and portable.  To the greatest
extent possible I avoid just giving direct access to the low-level C routines
such as fork, socket, setsockopt, dup2, bind, listen, waitpid, sigaction etc.
There are so many options and flags for those routines that I don't want to
reflect them all in Fexl.  My aim is to make a portable high-level layer that
does the right thing.  Let C do what it is good at and let Fexl do what it is
good at.

With that in mind I intend to improve the "spawn" function, which does not give
you the ability to interleave reading the stdout and stderr of the forked
process without fear of blocking.  I'll eventually have it call select(2) so
you can do that, and also detect the process exit at any time along the way
instead of forcing you to wait at the end.  I'm not yet sure how that interface
will look, but it will be something both portable and easy to use.

I added a library assoc.fxl which defines associative list functions.  I use
this extensively for the complex structured data I manage in hedge fund
accounting.  The test/b19.fxl file does some testing with 8192 random keys,
which is a bit atypical but it's something.

I added a library indent.fxl which outputs data values in a nested indented
format -- a "pretty-printer" so to speak.

I included a pair of test programs test/server.fxl and test/client.fxl to test
the client/server functions.  You can run "./fexl test/server.fxl start" to
start (or restart) the server.  You run "./fexl test/client.fxl" to connect to
the server.  This simple server merely increments each byte that you send it.
I'll eventually do a full web server, replacing my current server written in
Perl.  It will still run behind nginx which takes care of the SSL.

I added a "usleep" function which sleeps for a given number of microseconds.

I improved the file_content function (lib/list.fxl) so it reads 8K at a time
instead of individual bytes.

Here's a complete list of the new Fexl functions added and where they're
defined:

	usleep          type_run.c
	set_alarm       type_signal.c
	kill            type_run.c
	connect         type_run.c
	start_server    type_run.c
	stop_server     lib/run.fxl
	run_server      lib/run.fxl
	get_key         lib/assoc.fxl
	del_key         lib/assoc.fxl
	put_key         lib/assoc.fxl
	push_key        lib/assoc.fxl
	update_key      lib/assoc.fxl
	add_key         lib/assoc.fxl
	update_values   lib/assoc.fxl
	sort_pairs      lib/assoc.fxl
	stream_data     lib/indent.fxl
	put_data        lib/indent.fxl
	say_data        lib/indent.fxl
	trace_data      lib/indent.fxl