Fix readlink.
Now if you call readlink on a path that isn't a symbolic link it returns the
null string, as I originally intended.  Formerly it was doing a segmentation
fault, because at some point I switched to unsigned long in the str_new routine
and that caused a problem when readlink(2) returned -1.