Improve the sort function.
Formerly I used a quicksort algorithm which pivoted on the first item in the
list, but that gives terrible performance for lists that have large numbers of
items that are already in ascending or descending order.  Now I use a merge
sort which merges two adjacent segments together until there's only one left.
You can see this in the group_items function in lib/list.fxl.

Also, the group_items function is itself a very useful concept.  It takes a
list of items and a key extraction function and produces a list of groups.
Each group is a {key list} pair, where list is the list of all items with that
common key, in the same order as they appear in the original list (i.e. the
grouping is stable).  The groups are listed in ascending order of key.

The list of groups is in effect a partition of the original set, with the
equivalence class given by the key.  You can even look up entries in the
groups using the get_key function, which in effect gives you the inverse of the
key extraction function.

I also now call clear_free_list in type_fexl_benchmark, right before the
benchmark begins.  That gives you a better indication of how much memory is
used.  In particular it eliminates the negative memory usage you would see
when freeing memory during the benchmark.

I improved the comments on type_run_process and type_spawn in type_run.c.