Added the combine_keys function.
The group_items in lib/list.fxl is now written in terms of the more generally
useful combine_keys.  This allowed an optimization in the unique sort, which
now simply keeps only the first item instead of having to build the whole list
and then grab the first item when done.

I also added a test for group_items in test/b19.fxl.