Streamline some code.
Formerly eval called a pointer to a step function, and fexl_benchmark swapped
out the step function in order to count evaluation steps.  Now eval itself is a
pointer to a function, and fexl_benchmark swaps out the entire eval function.
This might yield some small performance gains during ordinary processing.  It
also provides the ability to do other kinds of specialized evaluation should
the need arise, such as tracing.  However, the main reason I did it was that I
like seeing the simple "g = f->T(f)" in the normal eval routine, instead of a
call to a one-line "step" routine.

I also unified the code for type_run_process and type_spawn, which now call a
common op_process routine.  That avoids the duplication of about 90 lines of
complicated code, putting it in one place now.