Fixed issue when building with older compiler.
In type_run.c on line 206, the older compiler produced an error message saying
that child_err might be uninitialized.  That's not exactly true, because it is
only uninitialized if catch_stderr is false, in which case I don't use it at
all.  But I went ahead and initialized it to 0 to avoid the message.