Make it more compatible with BSD.
I changed the headers in type_run.c so the socket code would build on BSD.

The lib_test script auto-locates "bash" now.  I noticed on a BSD machine that
bash was in /usr/local/bin/bash instead of /bin/bash.

I changed backtick notation to dollar-paren notation in lib_build.  This was
not necessary for BSD compatibility but I did it anyway.

Some tests involving processes and sockets are not working on BSD.  I'll have
to see why that is not portable.

I also noticed that "rand" is not compatible on BSD even with an identical
seed.  Eventually I'll port those tests to use an iterated hash instead of the
built-in rand so the output of things like stats.fxl will be compatible across
all machines.