Simplify and accelerate lambda abstraction and resolution of symbols.
This uses a new "struct form" to track the sorted list of distinct symbols in an
expression, along with the substitution pattern for each symbol.  This is very
quick and memory efficient during parsing, and now when we resolve a form we
only have to look up the definition of each distinct symbol once, instead of
looking it up for every occurrence of the symbol in the entire form.