Resolve numeric constants far more efficiently.
Formerly numeric symbols were treated just like any other name, and they would
be resolved by the context.  My technique of keeping all the unique symbols in
a form, along with their substitution patterns, is highly efficient -- however,
if you had a form with lots of distinct numbers, as might occur in a large data
file stored as Fexl, it was highly counterproductive, eating lots of memory and
time.

Now the parser recognizes numeric constants during the parse, with no need to
resolve them as symbols using a dynamic context function.  This is very
efficient in all cases.  However, it does imply that it is no longer possible
to use a number as a lambda symbol.  That is actually a good thing, because it
prevents semantic horrors such as the following expression returning true:

	(\0=1 eq 0 1)

Now if you try that, you get:

	Lambda name cannot be a number on line 1

I also took this opportunity to eliminate a deprecated feature in read_csv.fxl.
Formerly, when reading a CSV file, it would attempt to convert the items in a
row to numbers where possible -- for "convenience."  This feature was, from the
start, a monumentally Bad Idea.  Now it just leaves all the items as strings.